fix: tasks layout
