+++ src/components/moleculars/cards/ContributionCard/index.test.tsx
@@ -38,7 +38,7 @@ describe("ContributionCard with no variation", () => {
 
     clickOn("Donate");
 
-    expectPageToNavigateTo("/promoters/checkout", {
+    expectPageToNavigateTo("/promoters/recurrence", {
       search: "offer=1000&target=non_profit&target_id=1&currency=USD",
     });
   });
@@ -73,7 +73,7 @@ describe("ContributionCard with no variation", () => {
 
     clickOn("Donate");
 
-    expectPageToNavigateTo("/promoters/checkout", {
+    expectPageToNavigateTo("/promoters/recurrence", {
       search: "offer=1000&target=cause&target_id=1&currency=USD",
     });
   });
@@ -108,7 +108,7 @@ describe("ContributionCard with variation", () => {
 
     clickOn("Donate");
 
-    expectPageToNavigateTo("/promoters/checkout", {
+    expectPageToNavigateTo("/promoters/recurrence", {
       search: "offer=1000&target=non_profit&target_id=1&currency=USD",
     });
   });
@@ -144,7 +144,7 @@ describe("ContributionCard with variation", () => {
 
     clickOn("Donate");
 
-    expectPageToNavigateTo("/promoters/checkout", {
+    expectPageToNavigateTo("/promoters/recurrence", {
       search: "offer=1000&target=cause&target_id=1&currency=USD",
     });
   });
