+++ src/components/moleculars/cards/ContributionCard/index.tsx
@@ -89,7 +89,7 @@ function ContributionCard({
         currency,
       });
       navigateTo({
-        pathname: "/promoters/checkout",
+        pathname: "/promoters/recurrence",
         search: searchParams.toString(),
       });
     }
