+++ src/pages/users/ImpactPage/ImpactMenu/DirectSection/index.test.tsx
@@ -26,7 +26,7 @@ describe("Impact Page", () => {
         });
 
         it("should navigate to causes page", () => {
-          expectPageToNavigateTo("/promoters/checkout", {
+          expectPageToNavigateTo("/promoters/recurrence", {
             search: "offer=0&target=non_profit&target_id=&currency=USD",
           });
         });
